--- ethers-suave/src/const.ts
@@ -1,4 +1,0 @@
-
-export const CONFIDENTIAL_COMPUTE_RECORD_TYPE = '0x42'
-export const CONFIDENTIAL_COMPUTE_REQUEST_TYPE = '0x43'
-export const SUAVE_CHAIN_ID = 424242
