+++ hooks/useSuave.tsx
@@ -1,5 +1,5 @@
 import useBurnerWallet from "./useBurnerWallet"
-import { http } from '@flashbots/suave-viem';
+import { http, HttpTransport } from '@flashbots/suave-viem';
 import { getSuaveProvider, getSuaveWallet, type SuaveWallet } from '@flashbots/suave-viem/chains/utils';
 import { useEffect, useState } from "react";
 import { Chain, suaveToliman, suaveRigil } from "@flashbots/suave-viem/chains";
@@ -8,69 +8,16 @@ import { createPublicClient } from "@flashbots/suave-viem";
 export const toliman = suaveToliman
 export const rigil = suaveRigil
 
-// export const toliman: Chain = defineChain({
-//     id: 33626250,
-//     name: 'Toliman',
-//     network: 'toliman',
-//     nativeCurrency: {
-//         name: 'Toliman ETH',
-//         symbol: 'TEEth',
-//         decimals: 18
-//     },
-//     rpcUrls: {
-//         'default': {
-//             http: ['https://rpc.toliman.suave.flashbots.net'],
-//         },
-//         public: {
-//             http: ['https://rpc.toliman.suave.flashbots.net'],
-//         }
-//     },
-//     blockExplorers: {
-//         default: {
-//             name: 'Explorer',
-//             url: 'https://explorer.toliman.suave.flashbots.net/'
-//         },
-//     },
-// })
-
-// export const rigil: Chain = defineChain({
-//     id: 16813125,
-//     name: 'Rigil',
-//     network: 'rigil',
-//     nativeCurrency: {
-//         name: 'Rigil ETH',
-//         symbol: 'rigilETH',
-//         decimals: 18
-//     },
-//     rpcUrls: {
-//         'default': {
-//             http: ['https://rpc.rigil.suave.flashbots.net'],
-//         },
-//         public: {
-//             http: ['https://rpc.rigil.suave.flashbots.net'],
-//         }
-//     },
-//     blockExplorers: {
-//         default: {
-//             name: 'Explorer',
-//             url: 'https://explorer.rigil.suave.flashbots.net/'
-//         },
-//     },
-// })
-
 const useSuave = (chain: Chain = toliman) => {
+    const rpcUrl = chain.rpcUrls.public.http[0]
     const suaveClient = createPublicClient({
         chain: chain,
-        transport: http(chain.rpcUrls.default.http[0])
+        transport: http(rpcUrl)
     })
-
     const suaveProvider = getSuaveProvider(
-        http(chain.rpcUrls.default.http[0])
+        http(rpcUrl)
     )
-
     const { privateKey } = useBurnerWallet()
-
-    // @ts-expect-error    
     const [suaveBurnerWallet, setSuaveBurnerWallet] = useState<SuaveWallet<HttpTransport>>()
 
     useEffect(() => {
@@ -79,13 +26,13 @@ const useSuave = (chain: Chain = toliman) => {
             return
         }
         const newBurnerWallet = getSuaveWallet({
-            transport: http(chain.rpcUrls.default.http[0]),
+            transport: http(rpcUrl),
             privateKey: privateKey
         })
         setSuaveBurnerWallet(newBurnerWallet)
     }, [
         privateKey,
-        chain.rpcUrls
+        rpcUrl
     ])
 
     return {
